Added favourite and watch later lists
