+++ src/app/pages/now-playing-page/now-playing-page.component.ts
@@ -4,46 +4,23 @@ import { IMovie } from '../../models/IMovieCard';
 import { nowPlayingMovies } from '../../mock-data/mock-data';
 import { Router, RouterOutlet } from '@angular/router';
 import { SubHeaderComponent } from '../../components/sub-header/sub-header.component';
+import { BaseMoviesComponent } from '../../general/base-movies.component';
 
 @Component({
   selector: 'app-now-playing-page',
   standalone: true,
-  imports: [
-    MovieCardComponent,
-    RouterOutlet,
-    SubHeaderComponent,
-  ],
+  imports: [MovieCardComponent, RouterOutlet, SubHeaderComponent],
   templateUrl: './now-playing-page.component.html',
   styleUrl: './now-playing-page.component.scss',
 })
-export class NowPlayingPageComponent {
+export class NowPlayingPageComponent extends BaseMoviesComponent {
   movies: IMovie[] = nowPlayingMovies;
 
-  constructor(private router: Router) {}
+  constructor(private router: Router) {
+    super();
+  }
 
   navigateToDetail(id: number) {
     this.router.navigate(['now-playing', id]);
-  }
-
-  favouriteIds: number[] = [];
-  watchLaterIds: number[] = [];
-
-  addMovieToFavourite(movie: IMovie) {
-    const index = this.favouriteIds.indexOf(movie.id);
-    if (index > -1) {
-      this.favouriteIds.splice(index, 1);
-    } else {
-      this.favouriteIds.push(movie.id);
-    }
-    console.log(this.favouriteIds);
-  }
-
-  addMovieToWatchLater(movie: IMovie) {
-    const index = this.watchLaterIds.indexOf(movie.id);
-    if (index > -1) {
-      this.watchLaterIds.splice(index, 1);
-    } else {
-      this.watchLaterIds.push(movie.id);
-    }
   }
 }
