+++ src/app/pages/popular-page/popular-page.component.ts
@@ -3,18 +3,22 @@ import { MovieCardComponent } from '../../components/movie-card/movie-card.compo
 import { popularMovies } from '../../mock-data/mock-data';
 import { IMovie } from '../../models/IMovieCard';
 import { Router, RouterOutlet } from '@angular/router';
+import { BaseMoviesComponent } from '../../general/base-movies.component';
+import { SubHeaderComponent } from '../../components/sub-header/sub-header.component';
 
 @Component({
   selector: 'app-popular-page',
   standalone: true,
-  imports: [MovieCardComponent, RouterOutlet],
+  imports: [MovieCardComponent, RouterOutlet, SubHeaderComponent],
   templateUrl: './popular-page.component.html',
   styleUrl: './popular-page.component.scss',
 })
-export class PopularPageComponent {
+export class PopularPageComponent extends BaseMoviesComponent {
   movies: IMovie[] = popularMovies;
 
-  constructor(private router: Router) {}
+  constructor(private router: Router) {
+    super();
+  }
 
   navigateToDetail(id: number) {
     this.router.navigate(['popular', id]);
