+++ src/app/pages/top-rated-page/top-rated-page.component.html
@@ -1,11 +1,17 @@
 <div class="movies">
+  <app-sub-header
+    [favouriteMovieIds]="favouriteIds"
+    [watchLaterMovieIds]="watchLaterIds"
+  ></app-sub-header>
   <div class="movie-list">
     @for (movie of movies; track movie) {
-    <app-movie-card [movie]="movie" (viewDetail)="navigateToDetail($event)">
+    <app-movie-card
+      [movie]="movie"
+      (addFavourite)="addMovieToFavourite($event)"
+      (addWatchList)="addMovieToWatchLater($event)"
+      (viewDetail)="navigateToDetail($event)"
+    >
     </app-movie-card>
     }
-  </div>
-  <div>
-    <router-outlet name="list"></router-outlet>
   </div>
 </div>
