+++ src/app/pages/top-rated-page/top-rated-page.component.ts
@@ -3,18 +3,22 @@ import { MovieCardComponent } from '../../components/movie-card/movie-card.compo
 import { IMovie } from '../../models/IMovieCard';
 import { topRatedMovies } from '../../mock-data/mock-data';
 import { Router, RouterOutlet } from '@angular/router';
+import { BaseMoviesComponent } from '../../general/base-movies.component';
+import { SubHeaderComponent } from '../../components/sub-header/sub-header.component';
 
 @Component({
   selector: 'app-top-rated-page',
   standalone: true,
-  imports: [MovieCardComponent, RouterOutlet],
+  imports: [MovieCardComponent, RouterOutlet, SubHeaderComponent],
   templateUrl: './top-rated-page.component.html',
   styleUrl: './top-rated-page.component.scss',
 })
-export class TopRatedPageComponent {
+export class TopRatedPageComponent extends BaseMoviesComponent {
   movies: IMovie[] = topRatedMovies;
 
-  constructor(private router: Router) {}
+  constructor(private router: Router) {
+    super();
+  }
 
   navigateToDetail(id: number) {
     this.router.navigate(['top-rated', id]);
